+++ src/platform_impl/web/web_sys/event_handle.rs
@@ -1,7 +1,7 @@
 use wasm_bindgen::{prelude::Closure, JsCast};
-use web_sys::{AddEventListenerOptions, EventListenerOptions, EventTarget};
+use web_sys::{EventListenerOptions, EventTarget};
 
-pub(super) struct EventListenerHandle<T: ?Sized> {
+pub struct EventListenerHandle<T: ?Sized> {
     target: EventTarget,
     event_type: &'static str,
     listener: Closure<T>,
@@ -22,31 +22,6 @@ impl<T: ?Sized> EventListenerHandle<T> {
             event_type,
             listener,
             options: EventListenerOptions::new(),
-        }
-    }
-
-    pub fn with_options<U>(
-        target: &U,
-        event_type: &'static str,
-        listener: Closure<T>,
-        options: &AddEventListenerOptions,
-    ) -> Self
-    where
-        U: Clone + Into<EventTarget>,
-    {
-        let target = target.clone().into();
-        target
-            .add_event_listener_with_callback_and_add_event_listener_options(
-                event_type,
-                listener.as_ref().unchecked_ref(),
-                options,
-            )
-            .expect("Failed to add event listener");
-        EventListenerHandle {
-            target,
-            event_type,
-            listener,
-            options: options.clone().unchecked_into(),
         }
     }
 }